Gates introduction

+++ app/Http/Controllers/PostController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Http\Requests\StorePost;
 use App\Models\BlogPost;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class PostController extends Controller
 {
@@ -75,8 +76,16 @@ class PostController extends Controller
      */
     public function edit($id)
     {
+        $post = BlogPost::findOrFail($id);
+
+        $this->authorize('update-post', $post);
+        // if(Gate::denies('update-post', $post))
+        // {
+        //     abort(403, 'You Do Not Have The Permission To Edit This Blog Post');
+        // }
+
         return view('post.edit', [
-            'post' => BlogPost::findOrFail($id)
+            'post' => $post,
         ]);
     }
 
@@ -90,6 +99,13 @@ class PostController extends Controller
     public function update(StorePost $request, $id)
     {
         $post = BlogPost::findOrFail($id);
+
+        $this->authorize('update-post', $post);
+        // if(Gate::denies('update-post', $post))
+        // {
+        //     abort(403, 'You Do Not Have The Permission To Edit This Blog Post');
+        // }
+        
         $validated = $request->validated();
         
         $post->fill($validated);
@@ -111,6 +127,12 @@ class PostController extends Controller
     public function destroy($id)
     {
         $post = BlogPost::findOrFail($id);
+
+        $this->authorize('delete-post', $post);
+        // if(Gate::denies('delete-post', $post))
+        // {
+        //     abort(403, 'You Do Not Have The Permission To Delete This Blog Post');
+        // }
 
         $post->delete();
